agents/requirements.txt
crewai==0.95.0
python-dotenv>=1.0.0, <2.0.0
cdp-sdk>=0.15.0, <=0.17.0
web3>=7.8.0, <=8.0.0
